tapo-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
pydantic==2.7.4
tapo==0.8.12
httpx==0.27.0
